refactor(BA-6004): move Prometheus client into manager package
PrometheusClient, queriers, FixedQueryBuilder, metric/value types move from
common/clients/prometheus/ to manager/clients/prometheus/. preset.py stays in
common because common DTO validators import validate_query_template.
