+++ src/ai/backend/common/clients/prometheus/__init__.py
@@ -1,24 +1,8 @@
-from .client import PrometheusClient
-from .fixed_query_builder import (
-    ContainerLiveStatQueryBuilder,
-    ContainerMetricQueryBuilder,
-    LabelValuesQuery,
-)
 from .preset import LabelMatcher, LabelOperator, MetricPreset, validate_query_template
-from .querier import ContainerMetricQuerier, MetricQuerier
-from .types import MetricValue, ValueType
 
 __all__ = [
-    "ContainerLiveStatQueryBuilder",
-    "ContainerMetricQueryBuilder",
     "LabelMatcher",
     "LabelOperator",
-    "LabelValuesQuery",
-    "MetricValue",
-    "PrometheusClient",
     "MetricPreset",
-    "MetricQuerier",
-    "ContainerMetricQuerier",
-    "ValueType",
     "validate_query_template",
 ]
