+++ src/ai/backend/manager/clients/prometheus/__init__.py
@@ -1,0 +1,19 @@
+from .client import PrometheusClient
+from .fixed_query_builder import (
+    ContainerLiveStatQueryBuilder,
+    ContainerMetricQueryBuilder,
+    LabelValuesQuery,
+)
+from .querier import ContainerMetricQuerier, MetricQuerier
+from .types import MetricValue, ValueType
+
+__all__ = [
+    "ContainerLiveStatQueryBuilder",
+    "ContainerMetricQueryBuilder",
+    "ContainerMetricQuerier",
+    "LabelValuesQuery",
+    "MetricQuerier",
+    "MetricValue",
+    "PrometheusClient",
+    "ValueType",
+]
