+++ src/ai/backend/manager/clients/prometheus/querier.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from uuid import UUID
 
 from ai.backend.common.clients.prometheus.preset import LabelMatcher
-from ai.backend.common.clients.prometheus.types import ValueType
+from ai.backend.manager.clients.prometheus.types import ValueType
 
 
 class MetricQuerier(ABC):
