+++ src/ai/backend/manager/dependencies/domain/composer.py
@@ -17,7 +17,6 @@ from .repositories import RepositoriesDependency, RepositoriesInput
 from .services import ServicesContextDependency, ServicesInput
 
 if TYPE_CHECKING:
-    from ai.backend.common.clients.prometheus.client import PrometheusClient
     from ai.backend.common.clients.valkey_client.valkey_image.client import ValkeyImageClient
     from ai.backend.common.clients.valkey_client.valkey_live.client import ValkeyLiveClient
     from ai.backend.common.clients.valkey_client.valkey_schedule.client import (
@@ -25,6 +24,7 @@ if TYPE_CHECKING:
     )
     from ai.backend.common.clients.valkey_client.valkey_stat.client import ValkeyStatClient
     from ai.backend.common.etcd import AsyncEtcd
+    from ai.backend.manager.clients.prometheus.client import PrometheusClient
     from ai.backend.manager.config.provider import ManagerConfigProvider
     from ai.backend.manager.models.storage import StorageSessionManager
     from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
