+++ src/ai/backend/manager/services/metric/actions/live_stat.py
@@ -2,9 +2,9 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import override
 
-from ai.backend.common.clients.prometheus.metric_types import KernelLiveStatBatchResult
 from ai.backend.common.data.permission.types import EntityType
 from ai.backend.common.types import KernelId
+from ai.backend.manager.clients.prometheus.metric_types import KernelLiveStatBatchResult
 from ai.backend.manager.services.metric.actions.base import (
     QueryMetricAction,
     QueryMetricActionResult,
