+++ src/ai/backend/manager/services/metric/types.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
-from ai.backend.common.clients.prometheus.metric_types import (
+from ai.backend.common.types import BackendAISchema
+from ai.backend.manager.clients.prometheus.metric_types import (
     ContainerMetricOptionalLabel,
     ContainerMetricResponseInfo,
     ContainerMetricResult,
     MetricResultValue,
 )
-from ai.backend.common.clients.prometheus.types import ValueType
-from ai.backend.common.types import BackendAISchema
+from ai.backend.manager.clients.prometheus.types import ValueType
 
 __all__ = [
     "ContainerMetricOptionalLabel",
