+++ src/ai/backend/manager/services/prometheus_query_preset/service.py
@@ -1,8 +1,8 @@
 import logging
 
-from ai.backend.common.clients.prometheus.client import PrometheusClient
 from ai.backend.common.exception import PrometheusQueryPresetInvalidLabel
 from ai.backend.logging.utils import BraceStyleAdapter
+from ai.backend.manager.clients.prometheus.client import PrometheusClient
 from ai.backend.manager.data.prometheus_query_preset import (
     ExecutePresetOptions,
     PrometheusQueryPresetData,
