+++ tests/component/manager/clients/prometheus/BUILD
