+++ tests/component/manager/clients/prometheus/test_client_integration.py
@@ -8,14 +8,16 @@ import pytest
 
 from ai.backend.common.clients.http_client.client_pool import ClientPool, tcp_client_session_factory
 from ai.backend.common.clients.prometheus import (
-    ContainerLiveStatQueryBuilder,
-    ContainerMetricQueryBuilder,
     LabelMatcher,
     MetricPreset,
-    PrometheusClient,
 )
 from ai.backend.common.dto.clients.prometheus import PrometheusResponse, QueryTimeRange
 from ai.backend.common.exception import FailedToGetMetric
+from ai.backend.manager.clients.prometheus import (
+    ContainerLiveStatQueryBuilder,
+    ContainerMetricQueryBuilder,
+    PrometheusClient,
+)
 
 if TYPE_CHECKING:
     from ai.backend.testutils.bootstrap import PrometheusContainerFixture
