+++ tests/component/manager/clients/prometheus/test_sd_relabel.py
@@ -19,15 +19,17 @@ from ai.backend.common.clients.http_client.client_pool import (
     tcp_client_session_factory,
 )
 from ai.backend.common.clients.prometheus import (
-    ContainerLiveStatQueryBuilder,
-    ContainerMetricQueryBuilder,
     LabelMatcher,
     MetricPreset,
-    PrometheusClient,
 )
 from ai.backend.common.dto.clients.prometheus import PrometheusResponse
 from ai.backend.common.service_discovery.service_discovery import MODEL_SERVICE_GROUP
 from ai.backend.common.typed_validators import HostPortPair as HostPortPairModel
+from ai.backend.manager.clients.prometheus import (
+    ContainerLiveStatQueryBuilder,
+    ContainerMetricQueryBuilder,
+    PrometheusClient,
+)
 from ai.backend.testutils.pants import get_parallel_slot
 
 # ---------------------------------------------------------------------------
