+++ tests/unit/manager/clients/prometheus/BUILD
@@ -1,0 +1,1 @@
+python_tests(name="tests")
