+++ tests/unit/manager/repositories/prometheus_query_preset/test_prometheus_query_preset_options.py
@@ -11,8 +11,8 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from ai.backend.common.clients.prometheus.client import PrometheusClient
 from ai.backend.common.data.filter_specs import StringMatchSpec
+from ai.backend.manager.clients.prometheus.client import PrometheusClient
 from ai.backend.manager.models.prometheus_query_preset import PrometheusQueryPresetRow
 from ai.backend.manager.models.prometheus_query_preset.conditions import (
     PrometheusQueryPresetConditions,
