+++ tests/unit/manager/repositories/prometheus_query_preset/test_prometheus_query_preset_repository.py
@@ -9,7 +9,6 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from ai.backend.common.clients.prometheus.client import PrometheusClient
 from ai.backend.common.dto.clients.prometheus.response import (
     PrometheusQueryData,
     PrometheusResponse,
@@ -19,6 +18,7 @@ from ai.backend.common.exception import (
     PrometheusQueryEvaluationFailed,
     PrometheusQueryPresetNotFound,
 )
+from ai.backend.manager.clients.prometheus.client import PrometheusClient
 from ai.backend.manager.data.prometheus_query_preset import (
     PrometheusQueryPresetData,
 )
