+++ tests/unit/manager/services/prometheus_query_preset/test_prometheus_query_preset_service.py
@@ -12,7 +12,6 @@ from uuid import uuid4
 
 import pytest
 
-from ai.backend.common.clients.prometheus.client import PrometheusClient
 from ai.backend.common.dto.clients.prometheus.request import QueryTimeRange
 from ai.backend.common.dto.clients.prometheus.response import (
     PrometheusQueryData,
@@ -22,6 +21,7 @@ from ai.backend.common.exception import (
     PrometheusQueryPresetInvalidLabel,
     PrometheusQueryPresetNotFound,
 )
+from ai.backend.manager.clients.prometheus.client import PrometheusClient
 from ai.backend.manager.data.prometheus_query_preset import (
     ExecutePresetOptions,
     PrometheusQueryPresetData,
