+++ tests/unit/manager/services/utilization_metric/test_kernel_live_stat_batch.py
@@ -3,11 +3,11 @@ from uuid import UUID
 
 import pytest
 
-from ai.backend.common.clients.prometheus.metric_types import (
+from ai.backend.common.types import KernelId
+from ai.backend.manager.clients.prometheus.metric_types import (
     KernelLiveStatBatchResult,
     KernelLiveStatValues,
 )
-from ai.backend.common.types import KernelId
 from ai.backend.manager.services.metric.actions.live_stat import ContainerLiveStatAction
 from ai.backend.manager.services.metric.service import MetricService
